requirements.txt
Django==2.2
djangorestframework
djoser
django-filter
pyperclip
